+++ include/rocksdb/utilities/stackable_db.h
@@ -178,6 +178,13 @@ class StackableDB : public DB {
                                              import_options, metadata, handle);
   }
 
+  using DB::ClipColumnFamily;
+  virtual Status ClipColumnFamily(ColumnFamilyHandle* column_family,
+                                  const Slice& begin_key,
+                                  const Slice& end_key) override {
+    return db_->ClipColumnFamily(column_family, begin_key, end_key);
+  }
+
   using DB::VerifyFileChecksums;
   Status VerifyFileChecksums(const ReadOptions& read_opts) override {
     return db_->VerifyFileChecksums(read_opts);