+++ java/src/main/java/org/rocksdb/BlockBasedTableConfig.java
@@ -6,10 +6,10 @@ package org.rocksdb;
 
 /**
  * The config for plain table sst format.
- *
+ * <p>
  * BlockBasedTable is a RocksDB's default SST file format.
  */
-//TODO(AR) should be renamed BlockBasedTableOptions
+// TODO(AR) should be renamed BlockBasedTableOptions
 public class BlockBasedTableConfig extends TableFormatConfig {
 
   public BlockBasedTableConfig() {
@@ -243,7 +243,7 @@ public class BlockBasedTableConfig extends TableFormatConfig {
    * Disable block cache. If this is set to true,
    * then no block cache should be used, and the {@link #setBlockCache(Cache)}
    * should point to a {@code null} object.
-   *
+   * <p>
    * Default: false
    *
    * @param noBlockCache if use block cache
@@ -257,10 +257,10 @@ public class BlockBasedTableConfig extends TableFormatConfig {
   /**
    * Use the specified cache for blocks.
    * When not null this take precedence even if the user sets a block cache size.
-   *
+   * <p>
    * {@link org.rocksdb.Cache} should not be disposed before options instances
    * using this cache is disposed.
-   *
+   * <p>
    * {@link org.rocksdb.Cache} instance can be re-used in multiple options
    * instances.
    *
@@ -276,7 +276,7 @@ public class BlockBasedTableConfig extends TableFormatConfig {
 
   /**
    * Use the specified persistent cache.
-   *
+   * <p>
    * If {@code !null} use the specified cache for pages read from device,
    * otherwise no page cache is used.
    *
@@ -327,7 +327,7 @@ public class BlockBasedTableConfig extends TableFormatConfig {
    * is less than this specified number and adding a new record to the block
    * will exceed the configured block size, then this block will be closed and
    * the new record will be written to the next block.
-   *
+   * <p>
    * Default is 10.
    *
    * @param blockSizeDeviation the deviation to block size allowed
@@ -414,7 +414,7 @@ public class BlockBasedTableConfig extends TableFormatConfig {
   /**
    * Use partitioned full filters for each SST file. This option is incompatible
    * with block-based filters.
-   *
+   * <p>
    * Defaults to false.
    *
    * @param partitionFilters use partition filters.
@@ -428,7 +428,7 @@ public class BlockBasedTableConfig extends TableFormatConfig {
   /***
    * Option to generate Bloom filters that minimize memory
    * internal fragmentation.
-   *
+   * <p>
    * See {@link #setOptimizeFiltersForMemory(boolean)}.
    *
    * @return true if bloom filters are used to minimize memory internal
@@ -442,7 +442,7 @@ public class BlockBasedTableConfig extends TableFormatConfig {
   /**
    * Option to generate Bloom filters that minimize memory
    * internal fragmentation.
-   *
+   * <p>
    * When false, malloc_usable_size is not available, or format_version &lt; 5,
    * filters are generated without regard to internal fragmentation when
    * loaded into memory (historical behavior). When true (and
@@ -452,21 +452,21 @@ public class BlockBasedTableConfig extends TableFormatConfig {
    * the reading DB has the same memory allocation characteristics as the
    * generating DB. This option does not break forward or backward
    * compatibility.
-   *
+   * <p>
    * While individual filters will vary in bits/key and false positive rate
    * when setting is true, the implementation attempts to maintain a weighted
    * average FP rate for filters consistent with this option set to false.
-   *
+   * <p>
    * With Jemalloc for example, this setting is expected to save about 10% of
    * the memory footprint and block cache charge of filters, while increasing
    * disk usage of filters by about 1-2% due to encoding efficiency losses
    * with variance in bits/key.
-   *
+   * <p>
    * NOTE: Because some memory counted by block cache might be unmapped pages
    * within internal fragmentation, this option can increase observed RSS
    * memory usage. With {@link #cacheIndexAndFilterBlocks()} == true,
    * this option makes the block cache better at using space it is allowed.
-   *
+   * <p>
    * NOTE: Do not set to true if you do not trust malloc_usable_size. With
    * this option, RocksDB might access an allocated memory object beyond its
    * original size if malloc_usable_size says it is safe to do so. While this
@@ -495,9 +495,9 @@ public class BlockBasedTableConfig extends TableFormatConfig {
 
   /**
    * Use delta encoding to compress keys in blocks.
-   *
+   * <p>
    * NOTE: {@link ReadOptions#pinData()} requires this option to be disabled.
-   *
+   * <p>
    * Default: true
    *
    * @param useDeltaEncoding true to enable delta encoding
@@ -521,10 +521,10 @@ public class BlockBasedTableConfig extends TableFormatConfig {
 
   /**
    * Use the specified filter policy to reduce disk reads.
-   *
+   * <p>
    * {@link org.rocksdb.Filter} should not be closed before options instances
    * using this filter are closed.
-   *
+   * <p>
    * {@link org.rocksdb.Filter} instance can be re-used in multiple options
    * instances.
    *
@@ -576,7 +576,7 @@ public class BlockBasedTableConfig extends TableFormatConfig {
 
   /**
    * Returns true when compression verification is enabled.
-   *
+   * <p>
    * See {@link #setVerifyCompression(boolean)}.
    *
    * @return true if compression verification is enabled.
@@ -602,7 +602,7 @@ public class BlockBasedTableConfig extends TableFormatConfig {
 
   /**
    * Get the Read amplification bytes per-bit.
-   *
+   * <p>
    * See {@link #setReadAmpBytesPerBit(int)}.
    *
    * @return the bytes per-bit.
@@ -613,27 +613,27 @@ public class BlockBasedTableConfig extends TableFormatConfig {
 
   /**
    * Set the Read amplification bytes per-bit.
-   *
+   * <p>
    * If used, For every data block we load into memory, we will create a bitmap
    * of size ((block_size / `read_amp_bytes_per_bit`) / 8) bytes. This bitmap
    * will be used to figure out the percentage we actually read of the blocks.
-   *
+   * <p>
    * When this feature is used Tickers::READ_AMP_ESTIMATE_USEFUL_BYTES and
    * Tickers::READ_AMP_TOTAL_READ_BYTES can be used to calculate the
    * read amplification using this formula
    * (READ_AMP_TOTAL_READ_BYTES / READ_AMP_ESTIMATE_USEFUL_BYTES)
-   *
+   * <p>
    * value  =&gt;  memory usage (percentage of loaded blocks memory)
    * 1      =&gt;  12.50 %
    * 2      =&gt;  06.25 %
    * 4      =&gt;  03.12 %
    * 8      =&gt;  01.56 %
    * 16     =&gt;  00.78 %
-   *
+   * <p>
    * Note: This number must be a power of 2, if not it will be sanitized
    * to be the next lowest power of 2, for example a value of 7 will be
    * treated as 4, a value of 19 will be treated as 16.
-   *
+   * <p>
    * Default: 0 (disabled)
    *
    * @param readAmpBytesPerBit the bytes per-bit
@@ -699,7 +699,7 @@ public class BlockBasedTableConfig extends TableFormatConfig {
 
   /**
    * Determine if index compression is enabled.
-   *
+   * <p>
    * See {@link #setEnableIndexCompression(boolean)}.
    *
    * @return true if index compression is enabled, false otherwise
@@ -710,7 +710,7 @@ public class BlockBasedTableConfig extends TableFormatConfig {
 
   /**
    * Store index blocks on disk in compressed format.
-   *
+   * <p>
    * Changing this option to false  will avoid the overhead of decompression
    * if index blocks are evicted and read back.
    *